In 6.x, the various shminfo fields are int rather than u_long.
Former-commit-id: 0f434bc765b37ef8ae149373a08f64fb74760ff6

+++ usr.bin/ipcs/ipc.c
@@ -213,7 +213,7 @@ begin_define
 define|#
 directive|define
 name|SHMINFO_XVEC
-value|X(shmmax, sizeof(u_long))			\ 			X(shmmin, sizeof(u_long))			\ 			X(shmmni, sizeof(u_long))			\ 			X(shmseg, sizeof(u_long))			\ 			X(shmall, sizeof(u_long))
+value|X(shmmax, sizeof(int))				\ 			X(shmmin, sizeof(int))				\ 			X(shmmni, sizeof(int))				\ 			X(shmseg, sizeof(int))				\ 			X(shmall, sizeof(int))
 end_define
 
 begin_define
